EDA and inferential statistics
As per [article] (https://www.post-gazette.com/news/transportation/2018/12/02/Healthy-Ride-Pittsburgh-bike-sharing-program-adding-stations-new-neighborhoods/stories/201812020119) published in on DEC 3, 2018; there was a noticeable drop in ridership from 2016 to 2017. I would like to to test whether this hypothesis is commensurate with the data. My consolidated dataset has data from from 2015, Quarter 2 to 2018, Quarter 4. The samples are independent of each other and are random. I would be perfoming a two sample, one-tailed upper test Z test. I would be also be performing a two sample Bootstrap hypothesis test with null hypothesis being that there is no difference in means with regards to the average number of trip counts. Also included is Data Visualization, Exploration and outlier removals.